Add: machine learning class2

+++ Java/P2420.java
@@ -4,10 +4,10 @@ import java.util.Scanner;
 public class P2420 {
 
     public static void main(String[] args) {
-        int a=0, b=0;
+        long a=0, b=0;
         Scanner sc = new Scanner(System.in);
-        a = sc.nextInt();
-        b=sc.nextInt();
+        a = sc.nextLong();
+        b=sc.nextLong();
         if(a>=b)
         System.out.println(a-b);
         else
